fix: dumpしたときにファイルの存在確認をする

+++ lib/tolk/export.rb
@@ -9,6 +9,7 @@ module Tolk
     end
 
     def dump
+      return unless File.exist?("#{destination}/#{name}.yml")
       File.open("#{destination}/#{name}.yml", "w+") do |file|
         file.write(Tolk::YAML.dump(data))
       end
